pyproject.toml
[tool.poetry]
name = "scene_generation"
version = "0.1.0"
description = ""
authors = ["gorjanradevski <gorjanradevski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.3"
torchvision = "^0.4.2"
transformers = "^2.3"
jupyter = "^1.0"
tqdm = "^4.41"
nltk = "^3.4"
matplotlib = "^3.1"
Pillow = "=6.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
